+++ simplq/src/components/common/Button/index.jsx
@@ -1,18 +1,3 @@
-import React from 'react';
-import styles from './button.module.scss';
+import Button from './Button';
 
-const StandardButton = (props) => {
-  const { onClick } = props;
-  return (
-    <button
-      type="submit"
-      onClick={onClick}
-      className={props.outlined ? styles['standard-button-outlined'] : styles['standard-button']}
-    >
-      {props.icon ? <div className={styles['icon']}>{props.icon}</div> : null}
-      <div className={styles['text']}>{props.children}</div>
-    </button>
-  );
-};
-
-export default StandardButton;
+export default Button;
